# Enhance AI analysis with bundling support and improved logging