--- a/src/analyzer/analyzers/llm_analyzers/general.cr
+++ b/src/analyzer/analyzers/llm_analyzers/general.cr
@@ -8,20 +8,103 @@ module Analyzer::AI
     @llm_url : String
     @model : String
     @api_key : String?
+    @max_tokens : Int32
 
     def initialize(options : Hash(String, YAML::Any))
       super(options)
       @llm_url = options["ai_provider"].as_s
       @model = options["ai_model"].as_s
       @api_key = options["ai_key"].as_s
+      @max_tokens = LLM.get_max_tokens(@llm_url, @model)
     end
 
     def analyze
       client = LLM::General.new(@llm_url, @model, @api_key)
       target_paths = select_target_paths(client)
-      target_paths.each { |path| analyze_file(path, client) }
+
+      if target_paths.empty?
+        logger.warning "No files selected for AI analysis"
+        return @result
+      end
+
+      # Use the bundling approach if we have a token limit
+      logger.info "AI Analysis using #{@llm_url} with model #{@model} (max tokens: #{@max_tokens})"
+
+      if @max_tokens > 0 && target_paths.size > 5
+        analyze_with_bundling(target_paths, client)
+      else
+        target_paths.each { |path| analyze_file(path, client) }
+      end
+
       Fiber.yield
       @result
+    end
+
+    private def analyze_with_bundling(paths : Array(String), client : LLM::General)
+      files_to_bundle = [] of Tuple(String, String)
+      paths.each do |path|
+        next if File.directory?(path) || !File.exists?(path) || ignore_extensions.includes?(File.extname(path))
+        relative_path = get_relative_path(base_path, path)
+        content = File.read(path, encoding: "utf-8", invalid: :skip)
+        files_to_bundle << {relative_path, content}
+      end
+
+      bundles = LLM.bundle_files(files_to_bundle, @max_tokens)
+      channel = Channel(Tuple(String, Int32)).new
+
+      bundles.each_with_index do |bundle, index|
+        spawn do
+          bundle_content, token_count = bundle
+          logger.info "Processing bundle #{index + 1}/#{bundles.size} (#{token_count} tokens)"
+          process_bundle(bundle_content, client)
+          channel.send({bundle_content, token_count})
+        end
+      end
+
+      bundles.size.times { channel.receive }
+    end
+
+    private def process_bundle(bundle_content : String, client : LLM::General)
+      prompt = "#{LLM::BUNDLE_ANALYZE_PROMPT}\n#{bundle_content}"
+      response = client.request(prompt, LLM::ANALYZE_FORMAT)
+
+      logger.debug "Bundle analysis response:"
+      logger.debug_sub response
+
+      begin
+        response_json = JSON.parse(response.to_s)
+        response_json["endpoints"].as_a.each do |ep|
+          url = ep["url"].as_s
+          method = ep["method"].as_s
+
+          # Extract file path from the endpoint data if available
+          # or use a default path
+          path_info = if ep.as_h.has_key?("file") && !ep["file"].as_s.empty?
+                        PathInfo.new(ep["file"].as_s)
+                      else
+                        PathInfo.new("#{base_path}/ai_detected")
+                      end
+
+          params = ep["params"].as_a.map do |param|
+            p = param.as_h
+            name = p["name"].as_s
+            param_type = if p.has_key?("param_type")
+                           p["param_type"].as_s
+                         elsif p.has_key?("type")
+                           p["type"].as_s
+                         else
+                           ""
+                         end
+            value = p["value"]? ? p["value"].as_s : ""
+            Param.new(name, value, param_type)
+          end
+
+          details = Details.new(path_info)
+          @result << Endpoint.new(url, method, params, details)
+        end
+      rescue e : Exception
+        logger.warning "Error parsing bundle response: #{e.message}"
+      end
     end
 
     private def select_target_paths(client : LLM::General) : Array(String)
@@ -41,11 +124,13 @@ module Analyzer::AI
         rescue e : Exception
           logger.debug "Error parsing filter response: #{e.message}"
           # fallback: analyze all files
+          return Dir.glob("#{base_path}/**/*").reject { |p| File.directory?(p) || ignore_extensions.includes?(File.extname(p)) }
         end
       else
         logger.debug_sub "AI::Analyzing all files"
       end
-      Dir.glob("#{base_path}/**/*")
+
+      Dir.glob("#{base_path}/**/*").reject { |p| File.directory?(p) || ignore_extensions.includes?(File.extname(p)) }
     end
 
     private def analyze_file(path : String, client : LLM::General)