--- a/src/analyzer/analyzers/llm_analyzers/ollama.cr
+++ b/src/analyzer/analyzers/llm_analyzers/ollama.cr
@@ -7,19 +7,103 @@ module Analyzer::AI
   class Ollama < Analyzer
     @llm_url : String
     @model : String
+    @max_tokens : Int32
 
     def initialize(options : Hash(String, YAML::Any))
       super(options)
       @llm_url = options["ollama"].as_s
       @model = options["ollama_model"].as_s
+      @max_tokens = LLM.get_max_tokens("ollama", @model)
     end
 
     def analyze
       ollama = LLM::Ollama.new(@llm_url, @model)
       target_paths = select_target_paths(ollama)
-      target_paths.each { |path| analyze_file(path, ollama) }
+
+      if target_paths.empty?
+        logger.warning "No files selected for Ollama analysis"
+        return @result
+      end
+
+      # Use the bundling approach if we have a token limit
+      logger.info "Ollama Analysis using model #{@model} (max tokens: #{@max_tokens})"
+
+      if @max_tokens > 0
+        analyze_with_bundling(target_paths, ollama)
+      else
+        # Fallback to individual file analysis if we can't determine token limits
+        target_paths.each { |path| analyze_file(path, ollama) }
+      end
+
       Fiber.yield
       @result
+    end
+
+    private def analyze_with_bundling(paths : Array(String), client : LLM::Ollama)
+      files_to_bundle = [] of Tuple(String, String)
+      paths.each do |path|
+        next if File.directory?(path) || !File.exists?(path) || ignore_extensions.includes?(File.extname(path))
+        relative_path = get_relative_path(base_path, path)
+        content = File.read(path, encoding: "utf-8", invalid: :skip)
+        files_to_bundle << {relative_path, content}
+      end
+
+      bundles = LLM.bundle_files(files_to_bundle, @max_tokens)
+      channel = Channel(Tuple(String, Int32)).new
+
+      bundles.each_with_index do |bundle, index|
+        spawn do
+          bundle_content, token_count = bundle
+          logger.info "Processing bundle #{index + 1}/#{bundles.size} (#{token_count} tokens)"
+          process_bundle(bundle_content, client)
+          channel.send({bundle_content, token_count})
+        end
+      end
+
+      bundles.size.times { channel.receive }
+    end
+
+    private def process_bundle(bundle_content : String, ollama : LLM::Ollama)
+      prompt = "#{LLM::BUNDLE_ANALYZE_PROMPT}\n#{bundle_content}"
+      response = ollama.request(prompt, LLM::ANALYZE_FORMAT)
+
+      logger.debug "Bundle analysis response:"
+      logger.debug_sub response
+
+      begin
+        response_json = JSON.parse(response.to_s)
+        response_json["endpoints"].as_a.each do |ep|
+          url = ep["url"].as_s
+          method = ep["method"].as_s
+
+          # Extract file path from the endpoint data if available
+          # or use a default path
+          path_info = if ep.as_h.has_key?("file") && !ep["file"].as_s.empty?
+                        PathInfo.new(ep["file"].as_s)
+                      else
+                        PathInfo.new("#{base_path}/ollama_detected")
+                      end
+
+          params = ep["params"].as_a.map do |param|
+            p = param.as_h
+            name = p["name"].as_s
+            param_type = if p.has_key?("param_type")
+                           p["param_type"].as_s
+                         elsif p.has_key?("type")
+                           p["type"].as_s
+                         else
+                           ""
+                         end
+            value = p["value"]? ? p["value"].as_s : ""
+            Param.new(name, value, param_type)
+          end
+
+          details = Details.new(path_info)
+          @result << Endpoint.new(url, method, params, details)
+        end
+      rescue e : Exception
+        logger.warning "Error parsing bundle response: #{e.message}"
+      end
     end
 
     private def select_target_paths(ollama : LLM::Ollama) : Array(String)
@@ -38,12 +122,14 @@ module Analyzer::AI
           return filtered["files"].as_a.map(&.as_s)
         rescue e : Exception
           logger.debug "Error parsing filter response: #{e.message}"
-          # fallback: 분석대상 전체 파일
+          # fallback: analyze all files
+          return Dir.glob("#{base_path}/**/*").reject { |p| File.directory?(p) || ignore_extensions.includes?(File.extname(p)) }
         end
       else
         logger.debug_sub "Ollama::Analyzing all files"
       end
-      Dir.glob("#{base_path}/**/*")
+
+      Dir.glob("#{base_path}/**/*").reject { |p| File.directory?(p) || ignore_extensions.includes?(File.extname(p)) }
     end
 
     private def analyze_file(path : String, ollama : LLM::Ollama)