Introduce a bundling approach for file analysis in both General and Ollama analyzers to manage token limits effectively. Improve logging for empty target paths and analysis details to enhance debugging and user feedback. Update prompts for bundle analysis in the LLM module.