@hahwul 2025-04-06
| Project Size   | Before Optimization | After Optimization | Improvement         |
|----------------|---------------------|--------------------|---------------------|
| Small Project  | 3.7112 s            | 3.2663 s           | 0.4449 s (12.0%)    |
| Large Project  | 101.3066 s          | 29.21 s            | 72.0966 s (71.2%)   |